feat(BE): orderList 정리 완료 후에만 삭제 및 똑같은 값 tmpBook에 있을 경우 입력되지 않게 수정

+++ backend/src/main/java/com/pororo/docar/domain/cartBook/service/CartBookService.java
@@ -1,5 +1,6 @@
 package com.pororo.docar.domain.cartBook.service;
 
+import com.pororo.docar.common.exception.BadRequestException;
 import com.pororo.docar.common.exception.ResourceNotFoundException;
 import com.pororo.docar.domain.cartBook.dto.BookSetList;
 import com.pororo.docar.domain.cartBook.entity.CartBook;
@@ -27,7 +28,6 @@ public class CartBookService {
 //    public List<Long> orderList = tmpBookService.orderList;
     public static List<Long> orderList = new LinkedList<>();
 
-
     public List<BookSetList> getBooks() {
         List<BookSetList> list = new ArrayList<>();
         List<CartBook> bookList = cartBookRepository.findAll();
@@ -38,60 +38,68 @@ public class CartBookService {
         return list;
     }
 
-
-
     @Transactional
     public List<BookSetList> getBooksByBookshelf() {
+        List<CartBook> bookList = cartBookRepository.findAll();
         List<TmpBook> list = new ArrayList<>();
         orderList.add(1L);
         orderList.add(2L);
         orderList.add(3L);
         orderList.add(4L);
         orderList.add(5L);
-        Long idx = orderList.remove(0);
+        Long idx = orderList.get(0);
 
-        List<CartBook> bookList = cartBookRepository.findAll();
-        for (CartBook cartBook : bookList) {
-            if (cartBook.getBook().getBookshelf().getId() == idx) {
-                TmpBook tmpBook = TmpBook.builder()
-                        .floor(cartBook.getFloor())
-                        .site(cartBook.getSite())
-                        .book(cartBook.getBook())
-                        .build();
-                list.add(tmpBook);
+        if (!orderList.isEmpty()) {
+            for (CartBook cartBook : bookList) {
+                if (cartBook.getBook().getBookshelf().getId() == idx) {
+                    TmpBook tmpBook = TmpBook.builder()
+                            .floor(cartBook.getFloor())
+                            .site(cartBook.getSite())
+                            .book(cartBook.getBook())
+                            .build();
+                    if (!tmpBookRepository.existsById(tmpBook.getBook().getId())) {
+                        list.add(tmpBook);
+                    }
+                }
             }
-        }
-        tmpBookRepository.saveAll(list);
+            tmpBookRepository.saveAll(list);
 
-        List<BookSetList> setList = new ArrayList<>();
-        List<TmpBook> setBookList = tmpBookRepository.findAll();
+            List<BookSetList> setList = new ArrayList<>();
+            List<TmpBook> setBookList = tmpBookRepository.findAll();
 
-        for (TmpBook tmpBook : setBookList) {
-            CartBook cartBook = CartBook.builder()
-                    .floor(tmpBook.getFloor())
-                    .site(tmpBook.getSite())
-                    .book(tmpBook.getBook())
-                    .build();
-            System.out.println(cartBook);
-            setList.add(new BookSetList().entityToDto(cartBook));
+            for (TmpBook tmpBook : setBookList) {
+                CartBook cartBook = CartBook.builder()
+                        .floor(tmpBook.getFloor())
+                        .site(tmpBook.getSite())
+                        .book(tmpBook.getBook())
+                        .build();
+                setList.add(new BookSetList().entityToDto(cartBook));
+            }
+            return setList;
+        } else {
+            throw new BadRequestException("모든 책장 정리가 완료되었습니다.");
         }
-        return setList;
     }
 
     @Transactional
     public void deleteSetBooks() {
         List<TmpBook> setBookList = tmpBookRepository.findAll();
         List<CartBook> doneBookList = cartBookRepository.findAll();
+        Long idx = orderList.remove(0);
 
-        for (TmpBook tmpBook : setBookList) {
-            for (CartBook cartBook : doneBookList) {
-                Long id = tmpBook.getBook().getId();
-                if (id == cartBook.getBook().getId()) {
-                    cartBookRepository.delete(cartBook);
+        if (!setBookList.isEmpty()) {
+            for (TmpBook tmpBook : setBookList) {
+                for (CartBook cartBook : doneBookList) {
+                    Long id = tmpBook.getBook().getId();
+                    if (id == cartBook.getBook().getId()) {
+                        cartBookRepository.delete(cartBook);
+                    }
                 }
             }
+            tmpBookRepository.deleteAll();
+        } else {
+            throw new BadRequestException("정리할 책이 없습니다");
         }
-        tmpBookRepository.deleteAll();
     }
 
     @Transactional
@@ -111,8 +119,13 @@ public class CartBookService {
 
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-    public CartBook getBookById(Long id) {
+    public CartBook findById(Long id) {
         return cartBookRepository.findById(id)
+                .orElseThrow(() -> new ResourceNotFoundException("Id를 확인해주세요."));
+    }
+
+    public TmpBook getBookById(Long id) {
+        return tmpBookRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Id를 확인해주세요."));
     }
